Large overhaul to do 3 things: - Add the -DryRun parameter - Add a comment to the top of the scripts - Add quotes when referencing related scripts

+++ scripts/initialize/onetime-bashrc.ps1
@@ -1,32 +1,63 @@
-. $env:MyWindowsScripts\common\start-execution.ps1
+<#
+.SYNOPSIS
+  Configures bash_profile to read from our config location.
+
+.PARAMETER DryRun
+	Display what actions would be taken but don't actually do them.
+
+.LINK
+	https://github.com/HeyZiko/mywindows/
+#>
+
+param (
+  [alias("d")][switch]$DryRun
+)
+
+. "$env:MyWindowsScripts\common\start-execution.ps1"
 
 $newline = "`r`n"
 $fileToUpdate = ".bash_profile"
 $target = "~\$fileToUpdate";
 $source = '$MyWindowsConfig/.bashrc_common'
-$include = "$($newline)test -f $source && . $source$newline$newline"
+$include = "test -f $source && . $source"
 
-Write-White "This script makes the primary $fileToUpdate read from $source."
+Write-White "
+========
+This script makes the primary $fileToUpdate read from $source.
+"
+#. "MyWindows-Update-Config-File $($DryRun ? "-DryRun" : $null) -Target $target -IncludeLine $include"
 
-Test-MyWindowsConfig
+
+Test-MyWindowsConfig $($DryRun ? "-DryRun " : $null)
 
 Write-White "Checking for $target"
 if (-not (Test-Path $target)) {
-  Write-White "No $target found, creating it."
-  $include | Set-Content $target
+  Write-White "No $target found, $($DryRun ? "DRYRUN: Pretend " : $null)creating it."
+  if($DryRun) {
+    Write-White "DRYRUN: Would have written the following into $target :"
+    Write-Cyan $include
+   }
+   else {
+    $include | Set-Content $target
+   }
 }
 else {
   Write-White "Found $target. Checking to see if you already configured it."
-  $existingText = Select-String -Path $target -pattern $source
+  $existingText = Select-String -Path $target -pattern  [Regex]::Escape($source)
   if($existingText) {
     Write-White "Looks like you've already configured $target, nothing more to do here!"
   }
   else {
-    Write-White "You haven't configured $target yet. Updating it now."
+    Write-White "You haven't configured $target yet. $($DryRun ? "DRYRUN: Pretend " : $null)Updating it now."
     $include += ((Get-Content $target) -join "$newline")
-    $include | Set-Content $target
+    if($DryRun) {
+      Write-White "DRYRUN: Would have written the following into $target :"
+      Write-Cyan $include
+     }
+     else {
+      $include | Set-Content $target
+     }
   }
 }
-prompt
 
-. $env:MyWindowsScripts\common\end-execution.ps1
+. "$env:MyWindowsScripts\common\end-execution.ps1"
