+++ scripts/maintain/rerun-taskbar-shortcuts.ps1
@@ -1,21 +1,37 @@
-. $env:MyWindowsScripts\common\start-execution.ps1
+<#
+.SYNOPSIS
+  Configures bash_profile to read from our config location.
+
+.PARAMETER DryRun
+	Display what actions would be taken but don't actually do them.
+
+.LINK
+	https://github.com/HeyZiko/mywindows/
+#>
+
+. "$env:MyWindowsScripts\common\start-execution.ps1"
 
 
-Write-White "This script adds pre-determined taskbar shortcuts."
+$target = Join-Path $env:APPDATA "Microsoft\Internet Explorer\Quick Launch\User Pinned\TaskBar"
+$source = "$Env:CloudProfile\.windows\taskbar-shortcuts\";
+$shortcuts = Join-Path $source "*.lnk"
 
-Test-CloudProfile
+Write-White "
+========
+This script adds pre-built taskbar shortcuts.
+"
 
-$source = "$Env:CloudProfile\.windows\taskbar-shortcuts";
+Test-MyWindowsConfig $($DryRun ? "-DryRun " : $null)
+
 if (-not (Test-Path $source)) {
   Write-White("Taskbar shortcut folder is missing. Expected:
     $source");
-  break;
+  exit 1;
 }
-$target = Join-Path $env:APPDATA "Microsoft\Internet Explorer\Quick Launch\User Pinned\TaskBar"
-$shortcuts = Join-Path $source "*.lnk"
 
-Write-White "Copying the following shortcuts to $target`:`n$shortcuts"
-Copy-Item -Path $shortcuts -Destination $target;
-prompt
+Write-White "$($DryRun ? "DRYRUN: Pretend " : $null)Copying the following shortcuts to $target`:`n$shortcuts"
+if(-not $DryRun) {
+  Copy-Item -Path $shortcuts -Destination $target;
+}
 
-. $env:MyWindowsScripts\common\end-execution.ps1
+. "$env:MyWindowsScripts\common\end-execution.ps1"
